requirements.txt
async-timeout==4.0.2
attrs==22.1.0
autoit==0.2.6
bidict==0.22.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
contextlib2==21.6.0
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cvzone==1.5.6
cycler==0.11.0
DateTime==4.7
Deprecated==1.2.13
distlib==0.3.6
execnet==1.9.0
fifo==0.0.3
filelock==3.9.0
Flask==2.2.2
Flask-SocketIO==5.3.2
fonttools==4.34.4
idna==3.3
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
keyboard==0.13.5
kiwisolver==1.4.4
libusb==1.0.26b5
MarkupSafe==2.1.1
matplotlib==3.6.2
mock==4.0.3
numpy==1.23.1
openapi-codec==1.3.2
opencv-python==4.6.0.66
packaging==21.3
path==16.5.0
path.py==12.5.0
Pillow==9.2.0
pkg-about==1.0.7
platformdirs==3.1.0
pluggy==1.0.0
py==1.11.0
PyAutoIt==0.6.5
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyserial==3.5
pytest==7.1.3
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.2.1
pyusb==1.2.1
pywin32==304
redis==4.3.4
requests==2.28.1
screeninfo==0.8.1
six==1.16.0
subprocess.run==0.0.8
termcolor==2.0.1
tomli==2.0.1
twilio==7.13.0
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.20.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
zope.interface==5.5.0
